backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.8.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
redis[asyncio]>=5.0.0
celery[redis]>=5.4.0
slowapi>=0.1.9
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.23.0
pydantic-settings>=2.5.0
pydantic[email]>=2.8.0
python-multipart>=0.0.9
bcrypt==4.2.0
passlib==1.7.4
celery[redis]>=5.4.0
redis>=5.0.0
fastapi[all]  # for WebSockets
twilio  # for SMS (optional)
geopy  # for geo calculations
python-dateutil>=2.8.0